Updates to smartchart ui endpoints
