+++ src/responsemodels/__init__.py
@@ -1,0 +1,3 @@
+'''Intialization file for response models module'''
+import logging
+logger = logging.getLogger('rcapi.responsemodels')
