+++ src/routers/smartchartui.py
@@ -1,21 +1,9 @@
-import json
 import logging
 import os
-import typing
-from fastapi import APIRouter, Request, Response
+from fastapi import APIRouter
+from src.responsemodels.prettyjson import PrettyJSONResponse
 
 from src.util.fhirclient import FhirClient
-
-class PrettyJSONResponse(Response):
-    media_type = "application/json"
-    def render(self, content: typing.Any) -> bytes:
-        return json.dumps(
-            content,
-            ensure_ascii=False,
-            allow_nan=False,
-            indent=4,
-            separators=(", ", ": "),
-        ).encode("utf-8")
 
 
 logger = logging.getLogger('rcapi.routers.smartchartui')
