+++ startlocal.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+hypercorn main:app --reload
